% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IH.VerifyTimePoints.R
\name{.VerifyTimePoints}
\alias{.VerifyTimePoints}
\title{object, must be one of \{"quad", "uni", "exp"\} indicating the distribution
from which the time points are to be calculated. For character input,
input 'nTimes' must also be provided. If a numeric vector, the
time points to be used. If 0 is not the first value, it will be
concatenated by the software.}
\usage{
.VerifyTimePoints(timePoints, nTimes, ...)
}
\description{
object, must be one of \{"quad", "uni", "exp"\} indicating the distribution
from which the time points are to be calculated. For character input,
input 'nTimes' must also be provided. If a numeric vector, the
time points to be used. If 0 is not the first value, it will be
concatenated by the software.
}
